+++ Assets/RCore.SheetX/Editor/GoogleSheetHandler.cs
@@ -904,7 +904,7 @@ namespace RCore.SheetX.Editor
 			fileContent = fileContent.Replace("LOCALIZATION_FOLDER", m_settings.GetLocalizationFolder(out bool isAddressable));
 			fileContent = fileContent.Replace("IS_ADDRESSABLE", isAddressable.ToString().ToLower());
 			fileContent = SheetXHelper.AddNamespace(fileContent, m_settings.@namespace);
-			m_writer.Write(m_settings.constantsOutputFolder, $"{pFileName}.cs", fileContent, SheetXExportFileType.Localization);
+			m_writer.Write(m_settings.constantsOutputFolder, $"{pFileName}.cs", fileContent, SheetXExportFileType.LocalizationConstants);
 			m_writer.Info($"Exported {pFileName}.cs!");
 
 			//Write file localized text component
@@ -912,7 +912,7 @@ namespace RCore.SheetX.Editor
 					return;
 			fileContent = fileContent.Replace("LOCALIZATION_CLASS_NAME", pFileName);
 			fileContent = SheetXHelper.AddNamespace(fileContent, m_settings.@namespace);
-			m_writer.Write(m_settings.constantsOutputFolder, $"{pFileName}Text.cs", fileContent, SheetXExportFileType.Localization);
+			m_writer.Write(m_settings.constantsOutputFolder, $"{pFileName}Text.cs", fileContent, SheetXExportFileType.LocalizationComponent);
 			m_writer.Info($"Exported {pFileName}Text.cs!");
 		}